+++ src/Enum/Http/StatusCode.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Xen3r0\JiraApiClient\Enum\Http;
+
+/**
+ * @codeCoverageIgnore
+ */
+enum StatusCode: int
+{
+    case TooManyRequests = 429;
+}